fix(durability): widen st_nlink to u64 for musl portability
On glibc/x86_64, `stat.st_nlink` is already `u64`, but on some musl
targets (notably aarch64-unknown-linux-musl) it is `u32`. Assigning it
directly into the generation-root file-witness `links: u64` field
therefore fails to compile on those targets even though the witness
schema correctly stores link count as a 64-bit integer.

Use an explicit `u64::from(stat.st_nlink)` widening conversion so the
file witness builds cleanly across both libc widths without relying on
platform-specific type identity matching the witness field. Behavior is
unchanged on glibc; musl targets stop failing at the type boundary.

+++ crates/frankensearch-durability/src/generation_root.rs
@@ -404,7 +404,10 @@ fn file_witness(
         device: stat.st_dev,
         inode: stat.st_ino,
         mode: stat.st_mode,
-        links: stat.st_nlink,
+        // `st_nlink` is `u64` on glibc/x86_64 but `u32` on some musl targets
+        // (aarch64-unknown-linux-musl), so widen explicitly rather than
+        // relying on the platform-specific width matching the witness field.
+        links: u64::from(stat.st_nlink),
         uid: stat.st_uid,
         gid: stat.st_gid,
         size,
